backend/requirements.txt
# AttentionX - Python Dependencies

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0

# AI & ML
google-generativeai==0.8.0
openai-whisper==20231117

# Audio Analysis
librosa==0.10.2
soundfile==0.12.1
pydub==0.25.1

# Video Processing
moviepy==1.0.3

# Computer Vision
mediapipe==0.10.14
opencv-python-headless==4.10.0.84

# Utilities
numpy==1.26.4
aiofiles==24.1.0
python-dotenv==1.0.1
